Create MainActivity.kt
for whatever reason. this main file would not be pushed up to github. i created a file in github and copied the code into it to force it to work. if there are any issues with it starting up. that would be why because it works perfectly fine on my end.